% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{filter_donors}
\alias{filter_donors}
\title{Choose most informative donors for imputation}
\usage{
filter_donors(data, n_donors, n_info, freq)
}
\arguments{
\item{data}{the input file that was generated with \link{input}}

\item{n_donors}{the number of donors' haplotypes to choose}

\item{n_info}{the number of columns in the data file before the columns of the haplotypes}

\item{freq}{the maximal minor allele frequency in variants to consider when choosing the donors}
}
\value{
the data file with the chosen haplotypes
}
\description{
The donors are chosen based on K. Wasik et al, 2021
This function is used as a nested function in phcp_impute()
}
\references{
K. Wasik et al., Comparing low-pass sequencing and genotyping for trait mapping in pharmacogenetics. BMC genomics 22, 197 (2021).
}
\seealso{
\link{phcp_impute} which is the imputation function that uses this function

\link{input} that generates the data input for this function
}
