% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{phcp_impute}
\alias{phcp_impute}
\title{Imputation for pseudo-haplotypes}
\usage{
phcp_impute(
  haps,
  ancient_tped,
  genetic_map,
  chr,
  choose_donors = F,
  freq_donors = 0.5,
  variants
)
}
\arguments{
\item{haps}{a data frame of Phased reference genomes in .haps format, which is the output of SHAPEIT2.}

\item{ancient_tped}{a data frame of pseudo-haplotype to impute in Plink tped format.}

\item{genetic_map}{a data frame of genetic map for the reference genome with 3 columns: chromosome number, genetic distance in Morgans, physical position. The columns header should be: chr, genetic_distance, bp}

\item{chr}{the number of chromosome to impute (integer).}

\item{choose_donors}{the number of haplotypes to choose from the reference data for the imputation. If FALSE, all haplotypes in the data will be used.}

\item{freq_donors}{the maximal minor allele frequency in variants to consider when choosing the donors}

\item{variants}{a data frame with variants to impute . If the arguments is missing, all variants of the reference panel will be be imputed.}
}
\value{
A data frames with the posterior probabilities of the possible genotypes in the imputed variants.
}
\description{
This function perform imputation of pseudo-haplotypes. The imputation is based on Lee and Li & Stephens hidden Markov model, as implemented in ChromoPainter, but with the hidden state representing a pair of haplotypes from the reference panel.
}
\references{
N. Li, M. Stephens, Modeling linkage disequilibrium and identifying recombination hotspots using single-nucleotide polymorphism data. Genetics 165, 2213-2233 (2003)

D. J. Lawson, G. Hellenthal, S. Myers, D. Falush, Inference of population structure using dense haplotype data. PLoS Genet 8, e1002453 (2012)
}
